Addressed comments which resulted in a bit of refactoring
